requirements.txt
loguru==0.7.0
python-dotenv==1.0.0 
ffmpeg-python==0.2.0
aiofiles==23.2.1

numpy==1.26.3
playwright==1.44.0

Pillow==10.2.0
easyocr==1.7.1
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0 

aiohttp==3.9.1
telethon==1.28.5
websockets==11.0.3
beautifulsoup4==4.12.2

opentelemetry-api==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-instrumentation==0.34b0